Move `validate_block` from VMState to VM
